added command  for parsing the number of characters in a project
